exch: add [[fallthrough]] markers
CID 1439400 (#1 of 1): Missing break in switch (MISSING_BREAK)
unterminated_case: The case for value 269680898U is not terminated by a 'break' statement.
CID 1440675 (#1 of 1): Missing break in switch (MISSING_BREAK)
unterminated_case: The case for value 975175711U is not terminated by a 'break' statement.
CID 1441518 (#1 of 1): Missing break in switch (MISSING_BREAK)
unterminated_case: The case for value 975175711U is not terminated by a 'break' statement.
CID 1441652 (#1 of 1): Missing break in switch (MISSING_BREAK)
unterminated_case: The case for value 975175710U is not terminated by a 'break' statement.