build: add ldd check for mapi.so
mapi.so cannot be directly fed to ldd -r, because it relies on symbols
coming from an executable rather than a library. A fake library
providing the same symbol set could be used to generate a testmapi.la
from php_mapi sources and have it link to zendfake.la. zendfake.la
would need to be installed, else libtool won't make it shared. To
bypass that, make zendfake a program. The symbol resolution check will
thus not be carried out by lddcheck, but rather as part of the program
link stage.