exmdb_provider: fix resource leak in exmdb_server_set_message_instance_conflict
CID 1442778 (#4 of 4): Resource leak (RESOURCE_LEAK)
14. leaked_storage: Variable pattachments going out of scope leaks
the storage it points to.

Fix variable "pattachments" going out of scope, which leaks the
storage it points to. The newly created pattachments object would not
be associated with the message object if b_inconflict is true,
rendering most of the remaining function useless.