exmdb_provider: delete special folder name check
Special folders can be renamed (and, in fact, they are, for
localization reasons), so the special_name() function with only
English names makes zero sense.

create_folder_by_properties already checks that no folder name is
duplicate (within the hierarchy level that it is being placed in).

Special folders also cannot be deleted (barring manual database
surgery), so altogether, it is already ensured that no folder will
ever bear the same name that any special folder is currently using.

In summary, this removal fixes the problem that a German store with
"Postausgang"-named outbox rejects the creation of a regular folder
titled "Outbox".