emsmdb: always save new blank unmodified attachments if so requested
gromox-1.8-128-g52092e6d9 made it so that the attachment_object is
properly marked as b_touched when properties are set. However, when
no properties are set, b_touched remains =false, and the attachment
again is not saved.

message_object::save gets it right, testing for b_new also.
Do the same in attachment_object.