exmdb_provider: synthesize mandatory properties for recipients
When a recipient is copied from exmdb to another provider or store,
MSMAPI will reject recipient structures that are incomplete, failing
the copy.

This affects e.g. the Cached Mode mechanism of Outlook and CDFv2/.msg
file export of MFCMAPI, the latter of which would report an
MAPI_E_BAD_VALUE error. Trial and error has shown that four
properties need to be present to satisfy MSMAPI. In the documentation
to IMessage::ModifyRecipients, these four make an appearance again.

The following call chains have been observed (the broken recipient is
in an embedded message, so functions may appear twice in a chain):

For Outlook Cached Mode:

1. exmdb_server_get_message_instance_rcpts_all_proptags
2. exmdb_client_local::get_message_instance_rcpts_all_proptags
3. message_object_get_recipient_all_proptags
4. message_object::create
5. rop_openmessage

1. exmdb_server_get_message_instance_rcpts_all_proptags
2. exmdb_client_local::get_message_instance_rcpts_all_proptags
3. message_object_get_recipient_all_proptags
4. message_object::create
5. rop_syncimportmessagechange

1. message_get_message_rcpts
2. message_read_message
3. message_read_message
4. exmdb_server_read_message
5. exmdb_client_local::read_message
6. icsdownctx_object_write_message_change
7. icsdownctx_object_get_buffer_internal
8. icsdownctx_object::get_buffer
9. rop_fasttransfersourcegetbuffer

For MFCMAPI loading/exporting the main message:

1. instance_load_message
2. instance_load_message
3. exmdb_server_load_message_instance
4. exmdb_client_local::load_message_instance
5. message_object::create
6. rop_openmessage

For MFC opening the embedded message:

1. exmdb_server_get_message_instance_rcpts_all_proptags
2. exmdb_client_local::get_message_instance_rcpts_all_proptags
3. message_object_get_recipient_all_proptags
4. message_object::create
5. rop_openembeddedmessage

1. exmdb_server_get_message_instance_rcpts
2. exmdb_client_local::get_message_instance_rcpts
3. message_object::read_recipients
4. rop_openembeddedmessage