oxcmail: resolve a repeated condition in load_mime_skeleton
The function tries to locate some HTML property, and IPM.Note.SMIME messages,
should not be having such a property owing to OXOSMIME v13 §2.1.3.1.1.2 (pg 13)
point 5. Based on this finding, we conclude that dropping the redundant check
is not the proper resolution; instead, the condition is to be edited.

References: GXH-13, GXL-232