data: use BLOB instead of NONE column type for SQLite DBs
The "NONE" specifier implies BLOB type and NONE affinity, which is
undesirable, because it allows INSERT statements like

	INSERT INTO x VALUES ("+43","035")

to treat those values as a number. The fact that quotes were used
does nothing in terms of affinity, unfortunately. We have to use the
"BLOB" specifier to get BLOB type and BLOB affinity.

The current online documentaion says on the matter:

	(Historical note: The "BLOB" type affinity used to be called "NONE".
	But that term was easy to confuse with "no affinity" and so it was
	renamed.)

This change of semantics happened for sqlite 3.0.0, over 17 years ago.

References: GXL-228, GXL-235
Fixes: gromox-0~666