rebuild: move file renaming to before unload
Triggerung unload first is the wrong thing to do, because that leaves
a window of opportunity where the old database gets loaded again.

   | rebuild       | exmdb_provider      | zcore
-------------------------------------------------------
1. | sends UNLOAD  |                     |
2. |               | unload exch.sqlite3 |
3. |               |                     | user log in
4. |               | loads exch.sqlite3  |
5. | renames files |                     |
6. |               | old DB in use...    |

In addition, when http is not running, then gromox-rebuild cannot
contact http and would abort without moving new.sqlite3 into place.
However, when http is not running, renaming files is safe anyway.