oxcmail: stop exporting PR_REPLY_RECIPIENT_NAMES
Remove the broken analysis of PR_REPLY_RECIPIENT_NAMES. That property
is not a PT_MV_UNICODE but just PT_UNICODE, with recipients separated
by semicolon.

If a recipient contains a semicolon as part of their name, Outlook
does no quoting/escaping and it becomes impossible to tell which
semicolon(s) are meant as a list separator and which are part of the
name. That makes PR_REPLY_RECIPIENT_NAMES useless altogether.

Fixes: gromox-0~666