oxcmail: reparse PR_REPLY_RECIPIENT_ENTRIES with proper type
PR_REPLY_RECIPIENT_ENTRIES is not a collection of ONEOFFs, but a
collection of arbitrary entryids. Parse it properly.

If and when there is a ONEOFF in the list however, that ONEOFF
contains the display name. That way, the information previously
thought "lost" (PR_REPLY_RECIPIENT_NAMES) is present again.

Fixes: gromox-0~666