doc: delete scenario 2 from gromox-kdb2mt
The normal scenario is that a Gromox host connects to Kopano
databases and reads data, i.e. a "pull" model.

If network/firewall configurations prohibit connections this way but
still allow connections the other way, an administrator could
theoretically extract Kopano databases in a "push" fashion.

To that end, Gromox utilities would have to be installed on the
Kopano host. This is doable, but there may be significant manual work
involved, because the set of distribution packages may have no
overlap. For example, Gromox delivers packages for Debian 11 and
Kopano 8.7.25 delivers for just Debian 10 presently. This makes the
push approach rather inconvenient, and so it is removed from
documentation.