emsmdb, zcore: perform dot-stuffing ahead of SMTP DATA
Certain lines in a mail have special meaning to certain components in
the chain, such as lines beginning with "From " in the Unix mbox
format, or lines beginning with a "." in a SMTP DATA command.

When regular mail is prepared for export, MIME::write_content always
applies a transform, and that is base64 or, (since recently)
alternatively, quoted-printable. The base64 transform naturally
"defuses" lines since neither "From " nor "." are possible with the
base64 output alphabet. The QP encoder specifically looks for the
special pattern and force-encodes them. By doing all that, no SMTP
dot-stuffing would be needed.

However, S/MIME is not classified as regular mail and would be copied
verbatim to the SMTP channel, which is incorrect.

Change the send routines to explicitly do SMTP-style dot-stuffing.

References: GXF-30, GXL-84