oxcmail: always generate angle brackets for exported addresses
When e.g. PR_SENDER_EMAIL_ADDRESS is absent or empty for whatever
reason, oxcmail used to generate

	From: "The Name"

without an addr-spec. This is not within the RFC 5322 grammar really,
and is likely to be reparsed as a local-part with an unspecified
domain, rather than a display name. Emitting some (equally
ungrammatical) <> gives parsers a potential hint what we intended to
mean with the "" part.

References: GXF-389