nsp: trial replacing ecMAPIOOM with ecServerOOM
ecMAPIOOM is used e.g. to convey that the client-side MAPI layer has
no more resources, or that it hit a static limit (like the 8K size
limit on direct-get properties). Most of the ecMAPIOOM emitted
however are due to the memory allocation limits in the server, for
which there is ecServerOOM.