nsp: disable OneOff synthesis for non-existing GAL objects
When the NSPI ResolveNames RPC is called with an SMTP address for
which there is no match in the Global Address List, NSP normally
reports an "unresolvable" status, and the MAPI client-side
addressbook root implementation will subsequently generate a OneOff
entryid for the MAPI application with ten properties.

	(PR_ADDRTYPE, PR_DISPLAY_NAME, PR_DISPLAY_TYPE,
	PR_EMAIL_ADDRESS, PR_ENTRYID, PR_OBJECT_TYPE, PR_RECORD_KEY,
	PR_SEARCH_KEY, PR_SEND_INTERNET_ENCODING, PR_SEND_RICH_INFO)

Our NSP instead report "resolved" and provided a fake contact with a
OneOff entryid itself. This is the primary bug.

As a result, the EMSMDB connector presents the contact to the MAPI
application with 15 properties.

	(Above list plus PR_7BIT_DISPLAY_NAME, PR_AB_PROVIDERS,
	PR_DISPLAY_TYPE, PR_SMTP_ADDRESS,
	PR_TRANSMITTABLE_DISPLAY_NAME)

Our NSP module does not send along PR_SMTP_ADDRESS in the
ResolveNames response, which is the second bug.

This patch temporarily adds a configuration directive to control the
behavior of OneOff synthesis. Ultimately, synthesis should just be
deleted completely because it is wrong.

References: DESK-359