emsmdb: add reference tracking to object handles
OXCROPS v24 §3.1.5.3 requires that clients release handles in a
certain order. (But it does not specify what happens if they don't.)
Either way, there is no ordering imposed on the handles of folders
with regard to any other folder, including subfolders.

When creating a folder in a non-default store with an Outlook that
operates in cached mode, it issues the following ROP sequence, which
is in line with the OXCROPS requirements:

	ropOpenFolder(inbox) = 5
	ropCreateFolder(5, "subfolder") = 6
	ropRelease(5)
	ropSetProperties(6, ...) = EBADF

The current implementation fails here, which is incorrect.