imap: avoid hanging clients if response buffer is full
	00000009 FETCH 1:42471 (UID ENVELOPE BODY.PEEK[HEADER.FIELDS
	(Newsgroups Content-MD5 Content-Disposition Content-Language
	Content-Location resent-to resent-date resent-from resent-cc
	resent-subject List-Help List-Unsubscribe List-Subscribe List-Post
	List-Owner List-Archive Followup-To References)] INTERNALDATE
	RFC822.SIZE FLAGS)

When the response buffer runs out of room because there were too many
results from e.g. such a FETCH, no command result would be written
(since the buffer is full), causing the client to wait on the server
forever. Take an early exit and yield a command error if that
happens.