lib: switch STREAM::copyline to work with unsigned quantities
rd_total_pos, rd_block_pos, etc. are unsigned already, so do the same
for loop indices.

Try to quench cov-scan's misconceptions about conditions.

CID 1491246 (#1 of 1): Out-of-bounds access (OVERRUN)
CID 1491231 (#1 of 1): Out-of-bounds access (OVERRUN)
19. overrun-buffer-arg: Calling memcpy with pbuff and actual_size is suspicious
because of the very large index, 18446744073709551615. The index may be due to
a negative parameter being interpreted as unsigned.