php_mapi: silence cov-scan warning about foreign macros
CID 1490536 (#1 of 1): Macro compares unsigned to 0 (NO_EFFECT)
unsigned_compare: This greater-than-or-equal-to-zero comparison
of an unsigned value is always true. 0U <= 8U.

It is true that 0U <= 8U is a tautology, but this is the result
of static evaluation by the preprocessor and really should not
be flagged in the first place.