oxcical: fix RECURRENCE-ID emission based on undefined contents
The block about PidLidExceptionReplaceTime was rewritten in
gromox-1.28-8-g0e4b78a5e. It used to have just "pvalue", but that was
replaced with "num" and "lnum" variables. Rewriting "pvalue != NULL"
to "lnum != nullptr" not quite correct, because even if
lnum!=nullptr, num==nullptr is still a possibility.

When PidLidIsException is present but PidLidStartRecurrenceTime is
not, a bogus RECURRENCE-ID line could be emitted to iCal, due to
logically unset "itime".

Explicitly track the state of itime now.

Fixes: gromox-1.28-8-g0e4b78a5e