lib: rectify return value of html_to_plain_boring
When w3m is absent from the system and the to-text conversion
produced exactly 65001 bytes, callers like oxcmail.cpp:oxcmail_import
or exmdb_provider/instbody.cpp:instance_conv_textfromhigher may
erroneously believe that the output was already UTF-8 when this was
not guaranteed. Furthermore, when the conversion produced 0 bytes,
callers would believe that conversion failed.
Stop returning lengths. The length is implicit in @outbuf.

Fixes: gromox-0.15-8-g9e4ae3c9c
Fixes: gromox-0.26-71-ge9aca0333