lib: rectify broken boundary checks of utf*_count_codepoints
In gromox-0~666, function utf8_len() was added with "len < clen" as
terminating condition. clen is the number of UTF-8 encoded input
bytes, and len the number of UTF-8 codepoints. This comparison makes
no sense. In UTF-8 encoding, there are always less-or-equal
codepoints than bytes. The condition is a tautology. In
gromox-1.11-92-gffd0021a0, utf8_len was renamed to
utf8_count_codepoints. In gromox-1.35-4-gcc42ebf17,
utf8_count_codepoints was removed.

In gromox-1.13-71-g2e00d4bae, utf8_count_codepoints was replicated to
utf16_count_codepoints. The comparison of len and clen still makes no
sense. Even in UTF-16 encoding, there are always less-or-equal
codepoints than bytes. Only surrogates would cause one character to
use two codepoints, but that's still 4 UTF-8 bytes -> 2 UTF-16
codepoints.

gromox-1.35-4-gcc42ebf17 exposed the bug because it rewrote the loop
such that "len" no longer counted the number of codepoints, but the
number of UTF-16 bytes, which easily leads to early loop termination
due to "len < clen" being false. This would report a string to be
shorter than it really is, therefore not filtering it for the 0x8000
byte limit of the RPC protocol.

Fixes: gromox-0~666
Fixes: gromox-1.13-71-g2e00d4bae