Add files via upload
Within the zipped file, I have separated the files into two parts.
1. The POP3 server implementation
2. The OAUTH 2.0 implementation in QT.
Below, I will detail the working operation, but looking forward to a video call to explain the working principle.

For the Oauth2.0 implementation.
You will need the Casdoor server installed and running locally on your PC for the testing, the detailed installation
procedure is on the casdoor website.
For this particular zipped project, The Casdoor server acts as the authorization server, where
The login email and password will be verified. The server handles this verification.

The QT app through the casdoor SDK acts as the client and serves as the interface to all the different unit.(resource server, authorization server, user).
The pop3sever is the resource server where data is fetched after authorization.
The QT app is gui , when you build it on the QT IDE, You will get the signin page from the authorization server(Casdoor server),
through the client(QT app with Casdoor SDK). On successful login, the  Casdoor server sends the access token to the client (QT app with SDK).
The client will share this access token with the resource server which will also verify the access token before data is sent to the client.
SInce the pop3server is a console app. I have used the Casdoor server(Authorization server) as the resource server, it is allowed.
so the resource server will send the user information(data, resource) from the Casdoor server back to the client(QT app with casdoor SDK). So it will  be displayed.
on the QT app.

For the pop3server implementation
The POP3server is the resource server, and will verify the access token before providing resources . So, in my implementation
The pop3server when turned on, will first verify an access token sent from the client against a list in a text file and also
verify the email. before sending data to the client.
I designed a separate pop3 console client to replace the QT app, so that both the server and client will communicate on terminal.
The pop3 implementation uses the standard commands used for transmitting and receiving data.