imap: redo parts of ps_literal_processing
This processing code is bullshit.

Loop boundary: If the last character is ``{``, the buffer is
incomplete and we need to wait. If the last characters are ``{}``, we
can process it. Hence, the n-2 loop bound is too aggressive; it needs
to be n-1.

Closing brace: should of course not start at read_context[i] (the
open brace) but one after. Then, ``read_len-i-1`` also makes sense.

Early exit: If no ``}`` was found after ``{``, then that situation
won't change in the next iteration. Replace ``continue`` by
``break``.

newline_size: Scan length is too much since we skipped over the
introducer already.

Variable names: atrocious and nondescriptive.

Parse the {octet} specifier just with strtoul: Remove the stupid
extra buffer. There is a guaranteed nondigit character in read_buffer
that makes strtol terminate.

Fixes: gromox-0~666
Fixes: gromox-1.24-90-g4506547bf