all: police file creation modes
Most utilities default to using a 0666/0777 create mode to the
open/mkdir calls. Judging by some stackoverflow vibes, this choice
was made to facilitate system services readily reading arbitrary
files created in the system volume. And this makes sense; consider
e.g. when root creates /etc/hostname (root:root:rw-r--r--). Were the
editor to use 0660 instead, root would have to issue an extra chmod
call (this also gives an explanation why the contemporary standard
umask is not 022 but 002).

Most of the data Gromox creates however is inherently private and not
meant for consumption by any other service or process on the system.
Because /var/lib/gromox already is 0770, a change in default file
modes for objects below that directory is mostly cosmetical.

References: DESK-1353