midb: distribute FOLDER-TOUCH event on all notifications
Synchronize deleted mailbox to Thunderbird when they are deleted with
g-web.

This moves the code block that broadcasts "FOLDER-TOUCH" outside of
the switch block. When gromox-imap receives this event, it refreshes
the desired mailbox folders. (FOLDER-TOUCH {username} {foldername},
whereas {foldername} is "inbox", "outbox", "trash", ...)

This broadcasting only happened in case of new_mail. So when the mail
was deleted, moved or copied with g-web, it never notified
Thunderbird to refresh the folders.

References: GXL-419