exmdb_local: avoid moving partially-completed delivered mails to cache/ directory
When a file is put in /var/lib/gromox/queue/cache, re-delivery will
be attempted. In case of partial_completion, that is a bad decision
though. Return a different error code from the innards of exmdb_local
so the file ends up in /var/lib/gromox/queue/save instead.

References: DESK-1353