midb_agent: ignore P-DTLU response rows with empty digests
If midb cannot produce the digest (E-1252 error in the system log),
it may emit just ``{}`` as a digest string. When midb_agent receives
such a response, it would signal failure to imapd, which then
abandons the command:

	< A FETCH 2103:2105 (BODY)
	> NO internal server error, 3/0x40ffff

Now, it will just skip that message:

	< A FETCH 2103:2105 (BODY)
	> * 2103 FETCH (...)
	> * 2105 FETCH (...)
	> A OK FETCH completed

This works well with Alpine 2.26.