oxcmail: ignore empty field bodies in header lines
For headers like From & To, an empty field body is generally not
allowed per the syntax grammar, neither per RFC 5322 §3 nor §4. In
other words, it constitutes an undefined state, and Alpine (the MUA)
ignores lines without content.

For headers like Subject, an empty field body is legal. Alpine
ignores such lines as well, which seems a reasonable thing to do from
a pragmatic view (especially if the code responsible for it does not
discriminate based on the header field name).

This change implements that Alpine logic in the Gromox RFC 5322
reader.

References: https://github.com/grommunio/grommunio-sync/issues/9
References: GXL-472