ews: avoid indirect function calls
The use of ``(a >= 4 ? f1 : f2)(b)`` leads to an indirect call. This
is ugly for many reasons: it could potentially interfere with branch
prediction/be penalized under spectre mitigations, it inhibits
inlining by the compiler, and it also does not save any bytes.

	401053: ba 76 11 40 00  mov    edx,0x401176 <_Z5ffff1PKc>
	401058: 50              push   rax
	401059: b8 82 11 40 00  mov    eax,0x401182 <_Z5ffff2PKc>
	40105e: 48 0f 44 c2     cmove  rax,rdx
	401062: 48 8b 7e 08     mov    rdi,QWORD PTR [rsi+0x8]
	401066: ff d0           call   rax
	401068:

A simpler ``a >= 4 ? f1(b) : f2(b)``:

	401053: 48 8b 7e 08     mov    rdi,QWORD PTR [rsi+0x8]
	401057: 83 f8 04        cmp    eax,0x4
	40105a: 75 07           jne    401063 <main+0x13>
	40105c: e8 15 01 00 00  call   401176 <_Z5ffff1PKc>
	401061: eb 05           jmp    401068 <main+0x18>
	401063: e8 1a 01 00 00  call   401182 <_Z5ffff2PKc>
	401068: