exmdb_client: discard connections when EOF is detected
After an unsuccessful RPC (and the response delivered to the client),
the exmdb server may close the connection. The closing however is
only noticed by the client at the next call, where it is erroneously
interpreted as that second RPC failing.

Whenever the client draws a connection from the connpool, evaluate
whether the file descriptor is still usable, else ditch it.