exmdb: simplify copy_eml_ext and resolve cov-scan warnings
link() can yield ENOENT if either the source or the destination is
missing. The idea was to print ENOENT errors only when the
destination directory was absent. But since srcdir is the same as
dstdir, all that is completely unnecessary. The change happens to
fix a harmless TOCTOU warning.

CID 1533068: (#2 of 2): Time of check time of use (TOCTOU)
4. fs_check_call: Calling function stat to perform check on
old_ext.c_str().