exmdb: move ephemeral DB onto disk
Support sqlite3 which is compiled with -DSQLITE_OMIT_SHARED_CACHE.
This also helps reduce memory usage, as MAPI tables can potentially
be large (unorderly inbox with 80K messages, anyone?). It is expected
that the OS kernel keeps it in the page cache as far as memory is
available to do so.

The presence of the DB on disk also benefits debugging, as one can
now invoke /usr/bin/sqlite3 on it.