ldap_adaptor: replace error messages about Size Limit Exceeded
LDAP_SIZELIMIT_EXCEEDED is not only emitted when the server hits a
preconfigured server-side limit, but also when the client-side limit
is reached. Therefore, there was inconsistent logging of this
condition. A server with no sizelimit (and suitably questionable data
set to trigger the condition) would cause ldap_adaptor to log:

	ldap_adaptor: search for mail=a@b.de yielded more than one match

while a server with sizelimit 1 would make ldap_adaptor log:

	ldap_adaptor: search with base "o=da" filter "mail=a@b.de": Size limit exceeded

Rewrite some parts to always use the first variant.

References: DESK-1852