svc_loader: delete unused get_state_path service function
There is no code left which uses the ``get_state_path()`` dlfunc, so
delete it. Now, there is no code left which uses the ``state_path``
config directives, so delete them as well.

(The mailbox directory paths are stored in MySQL, the working
directory for gromox-delivery-queue is settable via the
``dequeue_path`` directive, the working dataset for gromox-timer is
settable via ``timer_state_path``.)