svc_loader: silently accept reregistration of the same service function addresses
With making _ZT symbols visible (export) in default.sym, our .so
files moved into a position where they are the first provider of a
given _ZT symbol, which makes the file prone to stay resident as the
linker joins _ZT definitions of other .so files to adhere to C++ ODR
rules.

Our plugins do not do service function unregistration (reverse of
service_register_service() function), nor does it seem worth
implementing: if the .so file sticks around, so can the module state
and svc_loader state. Still we need to account for multiple
PLUGIN_INIT calls.

Fixes: gromox-2.27-85-g90b7c3e27