oxcmail: rewrite select_parts for better multialternative selection
Because of the MAPI message model, we do not select an alternative
like an MUA; instead, we select multiple (ouch).

Alternative selection should exclude multipart/mixed;
in contrast to MUAs, we cannot meaningfully treat them as a usable alternative.

Alternative selection should prefer the higher numbered parts;
the logic for this was missing as well.

References: GXL-526, DESK-2315