midb: resolve "inside a readonly TXN" warnings during message deletion
An implicitly started transaction stays open until the cursors have
reached the end of the data set. Just like e.g. fread(3), it is not
sufficient to read the exact number of rows/bytes — it really takes
that final step()/fread() call for the end-of-data/end-of-file flag
to be set.

Fixes: gromox-2.17-23-g24e7d57eb
Fixes: gromox-2.17-117-g5fb5e9a20
References: GXL-524, GXF-1738, GXF-1747, GXF-1769, DESK-2352