alias_resolve: switch from explicit map locks to std::atomic
Improve performance of concurrent access. Since C++20 is now
guaranteed in the codebase, std::atomic<std::shared_ptr> can be used
instead of explicit mutex locking. This should improve performance
slightly.

(Citing https://stackoverflow.com/a/40227410: """wrap the shared_ptr
with a lock, but this solution is not so scalable under some
contention, and in a sense, loses the automatic feeling of the
standard shared pointer.""")

Fixes: gromox-1.11-77-g4f4bef351