midb: terminate folder name encoding at the network boundary
Since HEAD^, there are no more special folder names in midb.sqlite3,
and all folders are stored in hexbin string form.

This encoding is no longer necessary for the database. It is even
detrimental for case-insensitive searches, which would have to be
done with

	SELECT ... WHERE name=? OR LOWER(UNHEX(name))=LOWER(UNHEX(?))

At the network level, the use of hexbin meant that folder names with
spaces would be encoded to whitespace-free names. That is a nice
property, and can be retained. Choosing base64 makes the encoding is
a bit more compact, and it also makes leftover hexbin occurrences
standout in debugging.

References: DESK-2409