midb: stop yielding TRYCREATE everywhere
RFC 3501 specifies [TRYCREATE] only for APPEND, COPY and UID COPY.

RFC 6851 specifies [TRYCREATE] also for MOVE and UID MOVE.

We originally added TRYCREATE for SELECT (gromox-2.7-61-g6ea93e5ed),
so maybe we should keep the general idea of (RFC quote): """hint to
the client that the operation can succeed if the mailbox is first
created""".

Emitting TRYCREATE for M-DELE (M-RENF) is pointless, because the
(source) message won't be found even if the client tries to
(re-)create the folder.

Fixes: gromox-2.37-66-gfd5fe4d2d