ews: save datetime props for offset calculation
Some clients send the start and end names as local time, and the
timezone information in separate tags on an appointment update, e.g.

```xml
<t:SetItemField>
  <t:FieldURI FieldURI="calendar:Start"/>
    <t:CalendarItem>
      <t:Start>2025-02-20T11:30:00</t:Start>
    </t:CalendarItem>
  </t:SetItemField>
<t:SetItemField>
<!-- … -->
<t:SetItemField>
  <t:ExtendedFieldURI PropertyName="CalendarTimeZone"
  PropertySetId="A7B529B5-4B75-47A7-A24F-20743D6C55CD" PropertyType="String"/>
  <t:CalendarItem>
    <t:ExtendedProperty>
      <t:ExtendedFieldURI PropertyName="CalendarTimeZone"
      PropertySetId="A7B529B5-4B75-47A7-A24F-20743D6C55CD" PropertyType="String"/>
      <t:Value>Europe/Vienna</t:Value>
    </t:ExtendedProperty>
  </t:CalendarItem>
</t:SetItemField>
```

Save these properties into the Shape and calculate the UTC time from
the local time sent by the client and the offset of the timezone.