emsmdb: fix cov-scan warning about indefinite wait
CID 1558768: (#1 of 1): Indefinite wait (BAD_CHECK_OF_WAIT_COND)
dead_wait: A wait is performed without ensuring that the condition is not
already satisfied while holding lock g_cond_mutex. This can cause a deadlock if
the notification happens before the lock is acquired.

[I believe cov-scan warns on unconditional wait(), because the condition may
already be true and there is no CV signal to be expected anytime soon.
–jengelh]

Fixes: gromox-0~666