oxcical: deal with strange allday events that do not start/end on midnight
An allday appointment with proper TZ metadata but which does not obey
start/end time being midnight, e.g. when the appointment is moved
back by one second through direct MAPI property modification, is
being shown differently in OL vs. oxcical_export.

Implement something similar to OL's logic for oxcical.cpp by testing
if the majority of the appointment's time occurs mostly on the next
day. (Assuming most appointments are 24 hours long, that means
testing for starthour>=12. It is more difficult if the appointment is
on DST day, but we will ignore that problem.)

References: vcs/g-sync/136