exmdb: read multi-server map from SQL rather than exmdb_list.txt
This changes the exmdb server and client component to use the
homeserver from mariadb rather than exmdb_list.txt. Both parts need
to be done simultaneously, as client and server need to be in
agreement about a particular prefix in the exmdb_callid::connect
call. exmdb_list.txt is now obsolete.