midb: handle "INBOX" name with notif_folder_added
"INBOX" is a valid name for a non-inbox in MAPI, but special rules
apply for IMAP, so midb renames folders as they pass through. This
rename logic is applied when midb scans a mailbox initially, or when
an existing folder's name changes to "INBOX", but we doing this for
the create event triggered by [php_mapi]
mapi_folder_createfolder("INBOX").