exmdb: add cgkreset functionality
exchange.sqlite3 stores CNs not only in the PR_CHANGE_KEY property,
but also the folders.change_number and messages.change_number SQL
column. cgkrepair (as a Gromox/MAPI client) cannot reach that. This
makes a server operation necessary.

cgkreset is this new function. In contrast to cgkrepair, it does not
try to determine whether objects need fixing; it just bumps
everything for simplicity of implementation.