email_lib: delete unused itime_type enumerations
itime_type::floating was supposed to represent "DATE-TIME values that
are not bound to any time zone in particular" (RFC 5545 p.33), and
itime_type::local one that is bound to a zone. However, since struct
ical_time does not track the zone, it makes no sense to claim it is
bound to any.

Delete itime_type::unspec as well, because code treated it the same
as floating.

We never did anything with
Fixes: gromox-2.19-21-g9ae6489bd