ab_tree: restore pre-filtering of objects with AB_HIDE_FROM_GAL for NSP
ab_base::m_users always carries all users (no hiding), because
different users can have different cloak_bits.

zcore has knowledge about when a GAL container is to be opened, and
that it has to create a MAPI table for it. So, zcore will make a copy
of the corresponding ab_base user list, curate it (filter hidden
users), and then attach the list object to the MAPI table for later
cleanup.

Unlike zcore RPC or EMSMDB, the NSPI protocol is rather stateless. A
server does not know when the client-side MAPI implementation opens a
GAL container/creates a MAPI table. All we see are requests "read GAL
at x, for y rows". This means that ab_base needs to do the filtering
ahead of time and keep that state, since nsp/ has no (good) way to
keep it.

That filtered list for container 0 was lost in the ab_tree rewrite.
Reinstate it.

The filtered list for container ids !=0 was retained, but was not
filtered.

Fixes: gromox-2.40-23-g91fc5e139
References: GXL-611, DESK-3546