exmdb: add write_message_v3 EXRPC
The deliver_message RPC has a "digest" parameter with which a process
(gromox-delivery) can indicate to the Information Store (exmdb) that
an RFC5322 Internet Mail file is logically linked to a newly-created
MAPI message.

The write_message RPC is the "not delivery, just placement" companion
function, which is used by the import tools. Having a digest
parameter is equally valuable. This commit adds the parameter and its
handling.

References: GXH-155, GXF-2229