http: new implementation of mod_rewrite
The original documentation can be seen with ``git show 9f4de7ee5``.
It says that \N refers to capture substrings; this is in line with
admin expectations. But mod_rewrite does not produce that result *at
all*. The original rules

```
/OAB/oab.xml => \0/ews/oab.php
\(/Microsoft-Server-ActiveSync\) => \1/z-push/index.php
```

produce just ``/ews/oab.php`` and ``/z-push/index.php``,
respectively, and \0 or \1 translates to nothing.

Here is a new implementation, using a better substitution. There is
no more limitation on the URI length, number of capture groups, and
the stack usage is reduced too (hence marking as GXL-51).

Fixes: gromox-0~666