mapi_lib: use std::min+static_cast in place of std::min<T>
`std::min<T>(a, b)` is short, but if the type of `a` or `b` were to change and
become incompatible, the compiler would stay silent. Thus, using `std::min(a,
static_cast<T>(b))` like in the rest of the codebase is preferred as there is
still much refactoring expected generally.