exmdb: send mail independently of exchange_emsmdb
exmdb_provider used to grab the "ems_send_mail" symbol from
exchange_emsmdb. This only works if emsmdb is in the same process,
which is decidedly not the case for gromox-istore. All that
ems_send_mail bring to the table for exmdb was an implied outgoing
SMTP server. exmdb can just read the config file itself to obtain the
outgoing SMTP for use with the more general cu_send_mail routine.